WIP: prepare for conic
